requirements.txt
langchain
openai
python-dotenv
pinecone-client
